把我的奖品和我的元气值挪到了YQPoint_views
