+++ app/YQPoint_views.py
@@ -1,3 +1,4 @@
+from generic.models import YQPointRecord
 from app.views_dependency import *
 from app.models import (
     Prize,
@@ -18,8 +19,64 @@ from app.utils import (
 )
 
 __all__ = [
-    'showPools'
+    'myYQPoint', 'myPrize', 'showPools'
 ]
+
+
+@login_required(redirect_field_name="origin")
+@utils.check_user_access(redirect_url="/logout/")
+@log.except_captured(source='views[myYQPoint]', record_user=True)
+def myYQPoint(request: HttpRequest):
+    valid, user_type, html_display = utils.check_user_type(request.user)
+    # 获取可能的提示信息
+    my_messages.transfer_message_context(request.GET, html_display)
+
+    html_display.update(
+        YQPoint=request.user.YQpoint,
+    )
+ 
+    received_set = YQPointRecord.objects.filter(
+        user=request.user,
+    ).exclude(source_type=6).order_by("-time")
+
+    send_set = YQPointRecord.objects.filter(
+        user=request.user,
+        source_type=6,
+    ).order_by("-time")
+
+    # 新版侧边栏, 顶栏等的呈现，采用 bar_display, 必须放在render前最后一步
+    bar_display = utils.get_sidebar_and_navbar(request.user, "我的元气值")
+    return render(request, "myYQPoint.html", locals())
+
+
+@login_required(redirect_field_name="origin")
+@utils.check_user_access(redirect_url="/logout/")
+@log.except_captured(source='views[myPrize]', record_user=True)
+def myPrize(request: HttpRequest):
+    valid, user_type, html_display = utils.check_user_type(request.user)
+    # 获取可能的提示信息
+    my_messages.transfer_message_context(request.GET, html_display)
+
+    lottery_set = PoolRecord.objects.filter(
+        user=request.user,
+        pool__type=Pool.Type.LOTTERY,
+        status__in=[
+            PoolRecord.Status.LOTTERING, 
+            PoolRecord.Status.NOT_LUCKY,
+            PoolRecord.Status.UN_REDEEM],
+    ).order_by("-time")
+
+    exchange_set = PoolRecord.objects.filter(
+        user=request.user,
+        status__in=[
+            PoolRecord.Status.UN_REDEEM,
+            PoolRecord.Status.REDEEMED, 
+            PoolRecord.Status.OVERDUE],
+    ).order_by("-status", "-time")
+
+    # 新版侧边栏, 顶栏等的呈现，采用 bar_display, 必须放在render前最后一步
+    bar_display = utils.get_sidebar_and_navbar(request.user, "我的奖品")
+    return render(request, "myPrize.html", locals())
 
 
 @login_required(redirect_field_name="origin")
